Change in how we interpret equality
I am not treating a comparison of a scalar value (i.e. 0) 
equivalent to a Money class with the same amount (i.e. Money(0, 'USD'))

I would rather take a cautious approach with unit conversion and force the 
API user to be explicit. I could potentially be convinced otherwise, but for now I 
feel more comfortable with this approach.